backend/requirements.txt
fastapi
paho-mqtt
pyserial